==> Scripts/1 Data import & descriptive analysis.ipynb <==
# Contents
#### 1. Importing libraries
#### 2. Importing data
#### 3. Descriptive analysis

# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

path = r'/Users/zoegyftopoulos/Documents/Data Projects/07-2024 Instacart Basket Analysis'

df = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'), nrows=1000)

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'), usecols = vars_list)

# 3. Descriptive analysis

df.head(5)

df.prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'))

df.prods.head(20)

df.prods.tail(35)

df.prods.columns

df.prods.shape

df.prods.describe()

df.prods.dtypes
