==> Scripts/2 Data wrangling & subsetting.ipynb <==
# Contents
#### 1. Importing libraries
#### 2. Importing data
#### 3. Data wrangling
##### - Transposing data frame
##### - Creating a dictionary
#### 4. Exporting data

# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

path = r'/Users/zoegyftopoulos/Documents/Data Projects/07-2024 Instacart Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'))

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'))

# 3. Data wrangling

# Dropping eval_set column from orders.csv
df_ords.drop(columns = ['eval_set'])

df_ords.head()

# Check for missing values
df_ords['days_since_prior_order'].value_counts(dropna = False)

# Rename column
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

df_ords.describe()

# Change data type of order_id
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

### Transpose the departments data frame

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'departments.csv'))

df_dep.head()

# Transpose the data
df_dep.T

df_dep_t = df_dep.T

df_dep_t.head()

df_dep_t.reset_index()


new_header = df_dep_t.iloc[0]

new_header

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

df_dep_t_new.columns = new_header

df_dep_t_new

### Create a data dictionary with departments

data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19'))

### Subsetting

df_snacks =  df_prods[df_prods['department_id']==19]

df_prods['department_id']==19

df_prods[df_prods['department_id']==19]

df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

### Tasks

##### Find another identifier variable in the df_ords dataframe that doesn’t need to be included in the analysis as a numeric variable and change it to a suitable format

df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

##### Look for a variable in the df_ords dataframe with an unintuitive name and change its name without overwriting the dataframe

df_ords.rename(columns = {'order_hour_of_day' : 'order_time'}, inplace = True)

df_ords.head()

##### What is the busiest hour for placing orders? 

df_ords['order_time'].value_counts()

The busiest hour for placing orders is 10 am.

##### Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary

data_dict

The meaning behind a value of 4 in the "department_id" column is produce.

##### Create a subset containing breakfast item sales

df_breakfast =  df_prods[df_prods['department_id']==14]

df_breakfast.head()

###### Find all observations from the entire dataframe that include items for dinner parties from the following departments: alcohol, deli, beverages, and meat/seafood.

df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

df_dinner_parties.head()

df_dinner_parties.shape

##### Extract all information about the customer with a "user_id" of "1".

df_user1 = df_ords.loc[df_ords['user_id']=='1']
df_user1

df_user1.describe()

# 4. Exporting data

df_ords.to_csv(os.path.join(path, '02 Data','Prepared data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared data', 'departments_wrangled.csv'))

df_ords.shape
