==> Scripts/3 Consistency checks.ipynb <==
# Contents
#### 1. Importing libraries
#### 2. Importing data
#### 3. Consistency checks products
#### 4. Consistency checks orders
#### 5. Exporting data

# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

path = r'/Users/zoegyftopoulos/Documents/Data Projects/07-2024 Instacart Basket Analysis'

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'))

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_wrangled.csv'))

# 3. Consistency checks products.csv

df_prods.describe()

## Missing values

df_prods.isnull().sum()

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.shape

# Filter out rows with missing product names
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

## Duplicates

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

# Drop duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups

# 4. Consistency checks orders.csv

df_ords.describe()

'days_since_prior_order': does seem to have missing values, as it has a lower count than the other variables; the max value is 30, so maybe orders are capped at the one month mark; the min is 0, which may indicate a user's first order.
'order number': has a max of 100, which may show that order numbers are capped at 100.

#### Check for mixed-type data

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

No column has mixed data

#### Missing data

df_ords.isnull().sum()

Only the days_since_prior_order column has missing values, indicating that first users might be coded with 0.

# Subset the dataframe so that only nulls in days_since_prior_order are contained
df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_ords_nan

# Next, check whether the missind values in 'days_since_prior_oder' correspond to 'order_number' equal to 1

# Filter to contain all order_numbers equal to 1
df_first_ord = df_ords_nan[df_ords_nan['order_number'] == '1']

# Check if these have missing values in days_since_prior_order
df_first_ord_miss = df_first_ord['days_since_prior_order'].isna().all()

print(df_first_ord_miss)

# -> Every row where the order number equals 1 has a missing value for days_since_prior_order
# Therefore, I would not delete the missing values, as we can derive valuable information from these rows

#### Duplicates

df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

# There are no duplicates in the data set.

# 5. Exporting data

df_ords.to_csv(os.path.join(path, '02 Data','Prepared data', 'orders_checked.csv'))

df_prods.to_csv(os.path.join(path, '02 Data','Prepared data', 'products_checked.csv'))
