==> Scripts/4 Merging data sets.ipynb <==
# Contents
#### 1. Importing libraries
#### 2. Importing data
#### 3. Consistency checks
#### 4. Merging data sets
#### 5. Exporting data

# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

path = r'/Users/zoegyftopoulos/Documents/Data Projects/07-2024 Instacart Basket Analysis'

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders_products_prior.csv'))

df_ords_2 = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_checked.csv'))

df_ords_prior.head()

df_ords = df_ords_2.drop(columns=['Unnamed: 0.1','Unnamed: 0'])

df_ords.head()

df_ords_prior.shape

df_ords.shape

# 3. Consistency checks

### Check for mixed-type data

for col in df_ords_prior.columns.tolist():
  weird = (df_ords_prior[[col]].map(type) != df_ords_prior[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords_prior[weird]) > 0:
    print (col)

### Check for missing values

df_ords_prior.isnull().sum()

### Check for duplicates

df_dups = df_ords_prior[df_ords_prior.duplicated()]

df_dups

# 4. Merging data sets

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

# 5. Exporting data

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_combined.pkl'))

df_merged_large.shape
