==> Scripts/4.2 Merging data sets.ipynb <==
# Contents
#### 1. Importing libraries
#### 2. Importing data
#### 3. Merging data sets
#### 4. Exporting data

# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

path = r'/Users/zoegyftopoulos/Documents/Data Projects/07-2024 Instacart Basket Analysis'

df_orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_combined.pkl'))

df_orders_products_combined.shape

df_orders_products_combined.head()

df_prod_2 = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'products_checked.csv'))

df_prod_2.head()

df_prod = df_prod_2.drop(columns=['Unnamed: 0'])

df_prod.head()

# 3. Merging data sets

pd.merge(df_orders_products_combined,df_prod, on = ['product_id'], indicator = True)

# Drop  unnecessary columns 
df_ord_prod_combined = df_orders_products_combined.drop(['_merge'], axis=1)

pd.merge(df_ord_prod_combined,df_prod, on = ['product_id'], indicator = True)

df_merged = df_ord_prod_combined.merge(df_prod, on = ['product_id'], indicator = True)

df_merged.head()

# Confirm the results of the merge using the merge flag.
df_merged['_merge'].value_counts()

# 4. Exporting data

df_merged.to_pickle(os.path.join(path, '02 Data','Prepared data', 'ords_prods_merge.pkl'))
