==> Scripts/5 Deriving new variables.ipynb <==
# Contents
#### 1. Importing libraries
#### 2. Importing data
#### 3. Deriving new variables 
#### 4. Exporting data

# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

path = r'/Users/zoegyftopoulos/Documents/Data Projects/07-2024 Instacart Basket Analysis'

df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'ords_prods_merge.pkl'))

df_ords_prods_merge.head()

# Create a subset
df = df_ords_prods_merge[:1000000]

# 3. Deriving new variables

### Define a function to sort products into price ranges

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

df['price_range'] = df.apply(price_label, axis=1)

df.value_counts()

df['prices'].max()


### Use loc() & if-statements to sort products into price ranges

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df.value_counts()

# Repeat loc() function for entire dataframe

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merge.value_counts()

df_ords_prods_merge['orders_day_of_week'].value_counts(dropna = False)

### Create a for loop to summarize how busy each day of the week is

result = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

print(result)

df_ords_prods_merge['busiest_day'] = result

df_ords_prods_merge['busiest_day'].value_counts(dropna = False)

df_ords_prods_merge.head()

#### Tasks

#### Create a new variable for two busiest and slowest days of the week

result_2 = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_2.append("Busiest days")  
  elif value == 3 or value == 4:
    result_2.append("Least busy")
  else:
    result_2.append("Regularly busy")

print(result_2)

df_ords_prods_merge['busiest_days'] = result_2

df_ords_prods_merge['busiest_days'].value_counts(dropna = False)

df_ords_prods_merge.head()

The sum of the busiest days categories (32.434.212) match the busiest day categories as well as the total row count of the dataframe.

#### Create a new variable for busiest hours of the day

df_ords_prods_merge['order_time'].value_counts(dropna = False)

result_3 = []

for value in df_ords_prods_merge["order_time"]:
  if value in [10,11,14,15,13,12,16,9]:
    result_3.append("Most orders")  
  elif value in [17,8,18,19,20,7,21,22]:
    result_3.append("Average orders")
  else:
    result_3.append("Fewest orders")

print(result_3)

df_ords_prods_merge['busiest_period_of_day'] = result_3

df_ords_prods_merge['busiest_period_of_day'].value_counts(dropna = False)

# 4. Exporting data

df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_merged_new_v.pkl'))
