==> Scripts/6 Grouping & aggregating.ipynb <==
# Contents
#### 1. Importing libraries
#### 2. Importing data
#### 3. Grouping & aggregating
##### - Creating a loyalty flag
##### - Creating a spending flag
##### - Creating a frequency flag
#### 4. Exporting data

# 1. Importing libraries

import pandas as pd
import numpy as np
import os

# 2. Importing data

path = r'/Users/zoegyftopoulos/Documents/Data Projects/07-2024 Instacart Basket Analysis'

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_merged_new_v.pkl'))

# Create a subset
df = ords_prods_merge[:1000000]

df.shape

df.head()

# 3. Grouping & aggregating

df.groupby('product_name')

df.groupby('department_id').agg({'order_number': ['mean']})

df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

#### Aggregating with transform()

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merge.head()

ords_prods_merge.head(100)

pd.options.display.max_rows = None

#### Deriving columns with loc()

ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)

#### Tasks

# Find aggregated mean of the order_number column grouped by department_id
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

The result of the entire dataframe shows all of the departments, while the subset only showed some of them. 

#### Check differences in average product prices based on customer loyalty

# Check the basic statistics of the product prices for each loyalty category
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean','min','max']})

New customers spend around 3$ more than loyal customers. 

#### Target different types of spenders by creating a spending flag

ords_prods_merge['average_price'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

ords_prods_merge.head()

ords_prods_merge.loc[ords_prods_merge['average_price'] < 10, 'spending_flag'] = 'Low spender'

ords_prods_merge.loc[ords_prods_merge['average_price'] >= 10, 'spending_flag'] = 'High spender'

ords_prods_merge.head()

#### Determine frequent versus non-frequent customers

ords_prods_merge['median_days_betw_orders'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

ords_prods_merge.head()

ords_prods_merge.loc[ords_prods_merge['median_days_betw_orders'] > 20, 'order_frequency_flag'] = 'Non-frequent customer'

ords_prods_merge.loc[(ords_prods_merge['median_days_betw_orders'] > 10) & (ords_prods_merge['median_days_betw_orders'] <= 20), 'order_frequency_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['median_days_betw_orders'] <= 10, 'order_frequency_flag'] = 'Frequent customer'

ords_prods_merge.head()

ords_prods_merge['order_frequency_flag'].value_counts(dropna = False)

# 4. Exporting data

ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_merged_ga.pkl'))
