==> Scripts/7 Data wrangling & merging.ipynb <==
# Contents
#### 1. Importing libraries
#### 2. Importing data
#### 3. Wrangling data
#### 4. Consistency checks
#### 5. Importing data
#### 6. Merging data
#### 7. Exporting data

# 1. Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# 2. Importing data

path = r'/Users/zoegyftopoulos/Documents/Data Projects/07-2024 Instacart Basket Analysis'

df_customer = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'customers.csv'))

df_customer.shape

df_customer.head()

# 3. Wrangling data

### Renaming columns

df_customer.rename(columns={'First Name': 'first_name', 'Surnam': 'surname', 'Gender':'gender', 'STATE':'state', 'Age':'age'}, inplace=True)

df_customer.head()

### Check descriptive statistics

df_customer.describe()

# 4. Consistency checks

### Check for mixed-type data

for col in df_customer.columns.tolist():
  weird = (df_customer[[col]].map(type) != df_customer[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customer[weird]) > 0:
    print (col)

# Changing first_name data type to string
df_customer['first_name'] = df_customer['first_name'].astype('str')

type('first_name')

### Check for missing values

df_customer.isnull().sum()

### Check for duplicates

df_dups = df_customer[df_customer.duplicated()]

df_dups

# 5. Importing data

df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_merged_ga.pkl'))

df_ords_prods.head()

# 6. Merging data

# Check that key column user_id has the same data type in the two dfs
df_ords_prods.dtypes

df_customer.dtypes

# Drop _merge column from df_ords_prods
ords_prods = df_ords_prods.drop(columns=['_merge'])

# Merge the two dfs
ords_prods_cust = ords_prods.merge(df_customer, on = 'user_id', indicator = True)

ords_prods_cust.head()

# 7. Exporting data

ords_prods_cust.to_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_customers.pkl'))
